requirements.txt
Flask
pandas
numpy
scikit-learn
matplotlib
seaborn
pyyaml
requests
jupyter
notebook
joblib